+++ app/models/sshkeys.rb
@@ -14,6 +14,11 @@
 ## limitations under the License.
 ##
 class Sshkeys < BaseFascade
+  include SshHelper
+
+  
+  S3   = "s3"
+  RIAK = "riak"
 
   attr_reader :ssh_keys
 
@@ -30,7 +35,62 @@ class Sshkeys < BaseFascade
     return self
   end
 
+  #gemerate SSH key
+  def create(api_params, &block)
+
+    SSHKey.generate
+    key_name = params[:key_name] || current_user.first_name
+
+    @filename = key_name
+
+
+    wparams = { :name => key_name, :path => sshpub_loc }
+
+    @res_body = CreateSshKeys.perform(wparams, force_api[:email], force_api[:api_key])
+    if @res_body.class == Megam::Error
+      @res_msg = nil
+      @err_msg="Please contact #{ActionController::Base.helpers.link_to 'support !.', "http://support.megam.co/"}."
+      respond_to do |format|
+        format.js {
+          respond_with(@res_msg, @err_msg, @filename, :layout => !request.xhr? )
+        }
+      end
+    else
+      @err_msg = nil
+      options ={:email => current_user.email, :ssh_key_name => key_name, :ssh_private_key => k.private_key, :ssh_public_key => k.ssh_public_key }
+      upload = SshKey.perform(options, ssh_files_bucket)
+      if upload.class == Megam::Error
+        @res_msg = nil
+        @err_msg="SSH key #{key_name} creation failed."
+        @public_key = ""
+        respond_to do |format|
+          format.js {
+            respond_with(@res_msg, @err_msg, @filename, :layout => !request.xhr? )
+          }
+        end
+      else
+        @err_msg = nil
+        @res_msg = "SSH key #{key_name} created successfully"
+        @public_key = k.public_key
+        respond_to do |format|
+          format.js {
+            respond_with(@res_msg, @err_msg, @filename, :layout => !request.xhr? )
+          }
+        end
+      end
+  end
+  end
+
   private
+
+  #
+  def ssh_public_location
+    if Rails.configuration.storage_type == "s3"
+      sshpub_loc = vault_s3_url+"/"+current_user.email+"/"+key_name
+    else
+      sshpub_loc = current_user.email+"_"+key_name     #Riak changes
+    end
+  end
 
   #a private method that take the sshkeys collection and returns a hash
   def to_hash(ssh_keys_collection)
